requirements.txt
numpy
pickle
scikit-learn
torch
Pillow
facenet-pytorch
deepface
